requirements.txt
django>=4.2,<5.0
confluent-kafka>=2.3.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
